2131 refactor risk page component (#2180)
* Move logic/style for widgets from Risk.tsx to separate files.

* Simplify if statements

* Remove unused imports and arguments.

* Refactor class name declarations for consistency.

* Move color, pagination and other methods to utils.

* Refactor widget names.

* Reduce number of arguments passed to widgets.

* Apply consistent title capitalization.
