+++ frontend/src/pages/Risk/VulnerabilityPieChart.tsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import { useHistory } from 'react-router-dom';
+import { ResponsivePie } from '@nivo/pie';
+import { Paper } from '@mui/material';
+import { Point } from './Risk';
+import * as RiskStyles from './style';
+
+const VulnerabilityPieChart = (props: {
+  title: string;
+  data: Point[];
+  colors: any;
+  type: string;
+}) => {
+  const history = useHistory();
+  const { title, data, colors, type } = props;
+  const { cardRoot, cardSmall, header, chartSmall } = RiskStyles.classesRisk;
+  return (
+    <Paper elevation={0} className={cardRoot}>
+      <div className={cardSmall}>
+        <div className={header}>
+          <h2>{title}</h2>
+        </div>
+        <div className={chartSmall}>
+          <ResponsivePie
+            data={data as any}
+            innerRadius={0.5}
+            padAngle={0.7}
+            arcLabelsSkipAngle={10}
+            arcLinkLabelsSkipAngle={10}
+            colors={colors}
+            margin={{
+              left: 30,
+              right: 50,
+              top: 30,
+              bottom: 50
+            }}
+            onClick={(event) => {
+              if (type === 'vulns') {
+                history.push(`/inventory/vulnerabilities?severity=${event.id}`);
+              }
+            }}
+          />
+        </div>
+      </div>
+    </Paper>
+  );
+};
+export default VulnerabilityPieChart;
